# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/shareholders:
    get:
      tags:
        - Shareholders
      summary: List Shareholders
      description: Get a paginated list of Shareholders with optional filtering and sorting
      operationId: listShareholders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareholderList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Shareholders
      summary: Create Shareholder
      description: Create a new Shareholder
      operationId: createShareholder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareholderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shareholder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/share_transactions:
    get:
      tags:
        - ShareTransactions
      summary: List ShareTransactions
      description: Get a paginated list of ShareTransactions with optional filtering and sorting
      operationId: listShareTransactions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareTransactionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ShareTransactions
      summary: Create ShareTransaction
      description: Create a new ShareTransaction
      operationId: createShareTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareTransactionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Shareholder:
      type: object
      required:
        - id
        - company_id
        - name
        - holder_type
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        party_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        holder_type:
          type: string
        metadata:
          type: object
...
    CreateShareholderRequest:
      type: object
      required:
        - company_id
        - name
        - holder_type
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        party_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        holder_type:
          type: string
        metadata:
          type: object
...
    ShareholderList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Shareholder'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ShareTransaction:
      type: object
      required:
        - id
        - company_id
        - share_class_id
        - shareholder_id
        - txn_type
        - quantity
        - price_per_share
        - amount
        - txn_date
        - gl_posted
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        share_class_id:
          type: string
          format: uuid
        shareholder_id:
          type: string
          format: uuid
        txn_type:
          type: string
        quantity:
          type: string
        price_per_share:
          type: string
        amount:
          type: string
        counterparty_shareholder_id:
          nullable: true
          type: string
          format: uuid
        transfer_group_id:
          nullable: true
          type: string
          format: uuid
        posting_reference:
          nullable: true
          type: string
          maxLength: 80
        txn_date:
          type: string
          format: date
        gl_posted:
          type: boolean
        metadata:
          type: object
...
    CreateShareTransactionRequest:
      type: object
      required:
        - company_id
        - share_class_id
        - shareholder_id
        - txn_type
        - quantity
        - price_per_share
        - amount
        - txn_date
        - gl_posted
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        share_class_id:
          type: string
          format: uuid
        shareholder_id:
          type: string
          format: uuid
        txn_type:
          type: string
        quantity:
          type: string
        price_per_share:
          type: string
        amount:
          type: string
        counterparty_shareholder_id:
          nullable: true
          type: string
          format: uuid
        transfer_group_id:
          nullable: true
          type: string
          format: uuid
        posting_reference:
          nullable: true
          type: string
          maxLength: 80
        txn_date:
          type: string
          format: date
        gl_posted:
          type: boolean
        metadata:
          type: object
...
    ShareTransactionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ShareTransaction'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []